+++ packages/ui-components/src/radio/radio-group.tsx
@@ -173,8 +173,13 @@ function RadioGroup(props: RadioGroupProps): JSX.Element {
     const onChangeValue = (event: React.FormEvent<HTMLInputElement>): void => {
         const target = event.target as HTMLInputElement;
         const chosenIndex = Number(target.value);
-        setCurrentSelected(chosenIndex);
-        props.onChange?.(props.items[chosenIndex], event);
+        // controlled mode
+        if (props.value !== undefined) {
+            props.onChange?.(props.items[chosenIndex], event);
+            // uncontrolled mode
+        } else {
+            setCurrentSelected(chosenIndex);
+        }
     };
 
     useEffect(() => {