+++ packages/ui-components/docs/changelog.md
@@ -2,7 +2,7 @@
 title: Changelog
 ---
 
-## NEXT
+## 1.2.3
 
 -   Fixed an issue where when `RadioGroup`'s state was updated twice.